moved get route by id for post from home route to api post route and updated path
