+++ views/hobby.handlebars
@@ -33,7 +33,7 @@
         <button class="downvote">▼</button>
       </div>
       <div class="post-content">
-        <h2 class="post-title"><a href="/post/{{id}}">{{title}}</a></h2>
+        <h2 class="post-title"><a href="/api/post/{{id}}">{{title}}</a></h2>
         <p class="post-meta">Posted by u/{{user.username}} on {{created_at}}</p>
         <div class="post-body">
           {{!-- FUTURE DEVELOPMENT}}
